Split AssemblyResolver.cs into focused partial files
Refactor the 775-line AssemblyResolver.cs into 5 focused partial-class
files, each under 300 lines:

- AssemblyResolver.cs (~208 lines): Core lifecycle — fields, constants,
  constructor, finalizer, Dispose, InitializeLifetimeService,
  AddSearchDirectoriesFromRunSetting
- AssemblyResolver.Resolution.cs (~229 lines): Resolve pipeline —
  OnResolve, ReflectionOnlyOnResolve, OnResolveInternal,
  TryLoadFromCache, SafeLog
- AssemblyResolver.Search.cs (~274 lines): Search & load —
  SearchAssembly, SearchAndLoadAssembly, RequestedAssemblyNameMatchesFound,
  AddSubdirectories
- AssemblyResolver.FileSystem.cs (~70 lines): Virtual FS seams —
  DoesDirectoryExist, GetDirectories, DoesFileExist, LoadAssemblyFrom,
  ReflectionOnlyLoadAssemblyFrom
- AssemblyResolver.WinRT.cs (~32 lines): WinRT resolution —
  WindowsRuntimeMetadataReflectionOnlyNamespaceResolve (NETFRAMEWORK only)

No public/internal API changes. All #if guards preserved.

Fixes #9161
